First commit to the project
I'm not intending for this to be a particularly big project, mainly just
one that I can expand as and when I need to, as I need more freelancing
tools. My other goal is to have a project that I can keep my software
development knowledge fresh, try new things, and also keep up to date
with Zend Expressive.