Refactor how the three savings amounts are calculated
The internal workings were changed to avoid creating a series of
redundant calculators. Now, there's one percentage calculator, and one
that returns an object containing the three calculated amounts.

I'm not completely happy with how this works at the moment, but it's a
good start. I'm planning to progressively refine it.